+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## v0.24.3 (WIP)
 
+- Fixed incorrectly reported unique validator error for fields starting with name of another field (#6281; thanks @svobol13).
+
 - Reload the created/edited records data in the RecordsPicker UI.
 
 - Updated Go dependencies.